refactor: set up oclif 4 infrastructure
- Add oclif config section to package.json with topics and commands path
- Add oclif ^4.22.81 devDependency
- Simplify index.js to single re-export from @oclif/core
- Reorganize command files into nested directories for auto-discovery
  (e.g., mapping-state.js -> mapping/state.js)
- Update import paths in all moved files
