+++ commands/connect-events/stream/delete.js
@@ -1,13 +1,13 @@
-import * as api from '../../lib/connect/api.js'
+import * as api from '../../../lib/connect/api.js'
 import cli from '@heroku/heroku-cli-util'
 
 export default {
-  topic: 'connect:mapping',
+  topic: 'connect-events:stream',
   command: 'delete',
-  description: 'Delete an existing mapping',
-  help: 'Delete an existing mapping',
+  description: 'Delete an existing stream',
+  help: 'Delete an existing stream',
   args: [
-    { name: 'mapping' }
+    { name: 'stream' }
   ],
   flags: [
     { name: 'resource', description: 'specific connection resource name', hasValue: true },
@@ -18,11 +18,11 @@ export default {
   run: cli.command(async function (context, heroku) {
     await cli.confirmApp(context.app, context.flags.confirm)
 
-    await cli.action('deleting mapping', (async function () {
-      const connection = await api.withConnection(context, heroku)
+    await cli.action('deleting stream', (async function () {
+      const connection = await api.withConnection(context, heroku, api.ADDON_TYPE_EVENTS)
       context.region = connection.region_url
-      const mapping = await api.withMapping(connection, context.args.mapping)
-      const response = await api.request(context, 'DELETE', '/api/v3/mappings/' + mapping.id)
+      const stream = await api.withStream(context, connection, context.args.stream)
+      const response = await api.request(context, 'DELETE', `/api/v3/streams/${stream.id}`)
       if (response.status !== 204) {
         throw new Error(response.data.message || 'unknown error')
       }
