+++ commands/connect/mapping/delete.js
@@ -1,13 +1,13 @@
-import * as api from '../../lib/connect/api.js'
+import * as api from '../../../lib/connect/api.js'
 import cli from '@heroku/heroku-cli-util'
 
 export default {
-  topic: 'connect-events:stream',
+  topic: 'connect:mapping',
   command: 'delete',
-  description: 'Delete an existing stream',
-  help: 'Delete an existing stream',
+  description: 'Delete an existing mapping',
+  help: 'Delete an existing mapping',
   args: [
-    { name: 'stream' }
+    { name: 'mapping' }
   ],
   flags: [
     { name: 'resource', description: 'specific connection resource name', hasValue: true },
@@ -18,11 +18,11 @@ export default {
   run: cli.command(async function (context, heroku) {
     await cli.confirmApp(context.app, context.flags.confirm)
 
-    await cli.action('deleting stream', (async function () {
-      const connection = await api.withConnection(context, heroku, api.ADDON_TYPE_EVENTS)
+    await cli.action('deleting mapping', (async function () {
+      const connection = await api.withConnection(context, heroku)
       context.region = connection.region_url
-      const stream = await api.withStream(context, connection, context.args.stream)
-      const response = await api.request(context, 'DELETE', `/api/v3/streams/${stream.id}`)
+      const mapping = await api.withMapping(connection, context.args.mapping)
+      const response = await api.request(context, 'DELETE', '/api/v3/mappings/' + mapping.id)
       if (response.status !== 204) {
         throw new Error(response.data.message || 'unknown error')
       }
