+++ commands/connect/mapping/diagnose.js
@@ -1,6 +1,6 @@
-import * as api from '../../lib/connect/api.js'
+import * as api from '../../../lib/connect/api.js'
 import cli from '@heroku/heroku-cli-util'
-import diagnose from './diagnose.js'
+import diagnose from '../diagnose.js'
 
 export default {
   topic: 'connect:mapping',
