+++ commands/connect/mapping/write-errors.js
@@ -1,4 +1,4 @@
-import * as api from '../../lib/connect/api.js'
+import * as api from '../../../lib/connect/api.js'
 import cli from '@heroku/heroku-cli-util'
 
 export default {
