+++ package.json
@@ -15,16 +15,37 @@
   ],
   "license": "Apache-2.0",
   "type": "module",
-  "main": "index.js",
   "scripts": {
-    "test": "CONNECT_ADDON=connectqa mocha && standard",
-    "style-fix": "standard --fix"
+    "test": "CONNECT_ADDON=connectqa mocha",
+    "posttest": "standard",
+    "style-fix": "standard --fix",
+    "build": "oclif manifest"
   },
   "files": [
-    "/index.js",
-    "/lib",
-    "/commands"
+    "/commands",
+    "/lib"
   ],
+  "oclif": {
+    "bin": "heroku",
+    "commands": "./commands",
+    "topics": {
+      "connect": {
+        "description": "manage connections for Heroku Connect"
+      },
+      "connect:mapping": {
+        "description": "manage mappings on a Heroku Connect addon"
+      },
+      "connect:notifications": {
+        "description": "manage notifications on a Heroku Connect addon"
+      },
+      "connect-events": {
+        "description": "manage connections for Heroku Connect Events Pilot"
+      },
+      "connect-events:stream": {
+        "description": "manage mappings on a Heroku Connect Events Pilot addon"
+      }
+    }
+  },
   "dependencies": {
     "@heroku-cli/command": "^12.2.1",
     "@heroku/heroku-cli-util": "^8.0.15",
@@ -40,6 +61,7 @@
     "mocha": "^11.0.1",
     "mockdate": "^3.0.5",
     "nock": "^14.0.10",
+    "oclif": "^4.22.81",
     "sinon": "^21.0.0",
     "standard": "^17.1.0",
     "unexpected": "^13.2.1"
